Sync Health Center agent src Agent 	76865: Node: Agent should output a message on start (blocks Docs/Translation) - Use agent build date and version from Agent.cpp 	77983: OMR: Method names missing 	77983: OMR: Method names missing 	78078: OMR: Native memory being reported incorrectly 	78106: Node: Code cleanup - Use ToStdString() in spath() 	76192: Node: BVT / sanity type test for Jenkins builds - Add a version of the test for running with node-hc 	78061: OMR: Trace subscriber does not de-register on unload 	78013: OMR: memory leak in method lookup 	77943: OMR agent mishandles ~ in options 	77978: Last hcd file generated by headless is sometimes empty 	76865: Node: Agent should output a message on start (blocks Docs/Translation) - Add missing quotes and comma 	76865: Node: Agent should output a message on start (blocks Docs/Translation) - gyp changes for Node agent to set AGENT_VERSION in agent.cpp based on package.json 	76170: Win32 2003:  Initialization error for library j9jvmti27(-3): JVMJ9VM009E J9VMDllMain failed 	77392: 32-bit agent crashes on AIX if headless=on - Dynamically allocate stringbuffer 	77725: JVMTI memory leak in agent - correct call to complete() + fix compiler warnings 	76688: 'Monitored System' menu items enabled for Node agent

+++ runtime.tools/src/ibmras/common/util/LibraryUtils.cpp
@@ -91,7 +91,7 @@ std::string LibraryUtils::getLibraryDir(const std::string &library, const void* 
 #else
 	path = getLibraryLocation(func);
 #endif
-	int pos = path.find_last_of(PATHSEPARATOR);
+	size_t pos = path.find_last_of(PATHSEPARATOR);
 	if (pos != std::string::npos) {
 		path = path.substr(0, pos);
 	}
